Remove all the yaml stuff and the use of pragma once
With the yaml stuff we would never enter the LibreOffice
repository, since adding a dependency for a debug feature
is not really what we want. Pmd2raw and cppunit unit tests can
cater for this just fine.

As to the pragma once, we prefer to do it with defines. Also add
proper license header into all our files.

Change-Id: If11aec8a55a8d8dd7967d8a0ad228a73fcb0831f