Mark overriding member functions as 'override'
This does not affect the generated code (API or ABI), but:

1) Makes sure that in case the signature of one of these functions would
change by accident, the build breaks as the function no longer overrides
a base function.

2) Makes the code readable by explicitly marking all overriding
functions as 'override', where previously the reader had to read the
interface of the base class(es) as well to find out if the function is
virtual or not.

3) Since 'override' implies 'virtual', remove the 'virtual' keyword
where 'override' is present.

Change-Id: I04ce37eb53b8466be5257ab16d89a807ac00a14f